Moving all autohealing components to AutoHealing Module
