+++ src/app/auto-healing/autohealing-config-summary/autohealing-config-summary.component.css
