+++ src/app/auto-healing/autohealing.component.css
@@ -62,11 +62,11 @@
 }
 
 .activeAction {
-    color: #0585da;
+    color: #0058ad;
 }
 
 .image-btn {
-    color: #0585da;
+    color: #0058ad;
     border: none;
     font-size: 16px;
     cursor: pointer;
@@ -127,4 +127,14 @@ label {
 .topmenu-button{
     border-bottom-left-radius: 0px;
     border-bottom-right-radius: 0px;
+}
+
+.btn-primary {
+    background-color: #0058ad;
+    border-color: #0058ad;
+}
+
+.btn.btn-primary:disabled{
+    background-color: #0059ade7;
+    border-color: #0059ade7;
 }
