+++ src/app/availability/availability.routeconfig.ts
@@ -25,7 +25,7 @@ import { GenericDetectorComponent } from './generic-detector/generic-detector.co
 import { Injectable } from '@angular/core';
 import { GenericApiService } from '../shared/services/generic-api.service';
 import { TabTitleResolver } from '../shared/resolvers/tab-name.resolver';
-import { AutohealingComponent } from '../shared/components/autohealing/autohealing.component';
+import { AutohealingComponent } from '../auto-healing/autohealing.component';
 
 
 const _siteResourceUrl: string = 'subscriptions/:subscriptionid/resourcegroups/:resourcegroup/sites/:sitename';
